LayoutTests/webgl/2.0.0/conformance/limits/gl-line-width-expected.txt

This test runs the WebGL Test listed below in an iframe and reports PASS or FAIL.

Test: ../../resources/webgl_test_files/conformance/limits/gl-line-width.html
[ 1: PASS ] getError was expected value: NO_ERROR : Should be no errors.
[ 2: PASS ] getError was expected value: INVALID_VALUE : Should not be able to set lineWidth to zero
[ 3: PASS ] gl.getParameter(gl.LINE_WIDTH) is curLineWidth
[ 4: PASS ] getError was expected value: INVALID_VALUE : Should not be able to set lineWidth to a negative value
[ 5: PASS ] gl.getParameter(gl.LINE_WIDTH) is curLineWidth
[ 6: FAIL ] getError expected: INVALID_VALUE. Was NO_ERROR : Should not be able to set lineWidth to NaN
[ 7: FAIL ] gl.getParameter(gl.LINE_WIDTH) should be 1. Was NaN.
[ 8: PASS ] getError was expected value: NO_ERROR : Should be able to set lineWidth to the max supported value
[ 9: PASS ] gl.getParameter(gl.LINE_WIDTH) is lineWidthRange[1]
[ 10: PASS ] getError was expected value: NO_ERROR : Should be able to set lineWidth to the min supported value
[ 11: PASS ] gl.getParameter(gl.LINE_WIDTH) is lineWidthRange[0]
[ 12: PASS ] getError was expected value: NO_ERROR : Should be able to set lineWidth to the above max supported value
[ 13: PASS ] gl.getParameter(gl.LINE_WIDTH) is lineWidthRange[1]+1
[ 14: PASS ] successfullyParsed is true
[ FAIL ] 2 failures reported
